Add files via upload
is a registration system in php with Mysql database. Soon we will have an update on the budget generation part. Unzip the test file in your environment.
The database password and username are root. You can test in your environment or, within the .env file, change this data and do as you wish.